requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-static==0.0.6
Django==2.2.9
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.9
oauthlib==3.1.0
pipenv==2018.11.26
pipfile==0.0.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
toml==0.10.0
urllib3==1.25.9
virtualenv==16.7.9
virtualenv-clone==0.5.3
whitenoise==5.0.1
